Make singlecritter a fallback event (#3603)
Recurrence delays are global, so if enough events happen (e.g if admins addgamerule RSES 30 times) then they will keep rolling rarer and rarer events because all the less rare events will be out of the pool. This will keep singlecritter in the pool so there is always an alternative.