The Modest Grenade Resprite (#3612)
* The Whole Shebang

I should be doing smaller commits so my work gets actually backed up for once but in my defense most of this was just art that I was working on outside of the dev environment entirely.

* This doesn't fix the foam dart grenade

but it did fix the same problem I was having with the clusterbang so I don't know at this point.

* Actually I did fix them

the foam darts work good now

* I fixed modular grenades

also went a bit further and did new sprites for payloads and triggers too okay I love you bye-bye

* I put the smalled the metalfoam and bigged the clusterbang

All sprites on the metalfoam. Clusterbang gets the item component to change its inventory size, but its item size is kept at small so it fits in pockets and everything still.

* oops

oops